feat: Update dependencies and add lazy loading for native ads and images
Update dependencies in pubspec.yaml and pubspec.lock files to include the latest versions. Add lazy loading functionality for native ads in the live tab of the dashboard screen and in the live details screen. Also, implement lazy loading for images in the series details screen. These changes improve performance by loading content only when it is visible on the screen, reducing initial load time and enhancing the user experience.