feat: Update dependencies and add in-app update and social media sharing
Update dependencies in pubspec.yaml and pubspec.lock files to include the latest versions. Add the in_app_update package to enable in-app updates for the app. Also, integrate the flutter_share package to enable social media sharing functionality. These changes enhance the app's functionality and keep it up-to-date with the latest features and improvements.