feat: Update LiveDetails builder to include 'fromScorer' parameter
Update the builder function in the LiveDetails class to include a new 'fromScorer' parameter. This parameter is used to determine whether the LiveDetails screen is being accessed from the scorer screen or not. By adding this parameter, we can customize the behavior of the LiveDetails screen based on the source of navigation. This change improves the navigation flow and enhances the user experience.